game-flow: merge M_HandleEnableSunset to trx
