+++ src/libtrx/include/libtrx/game/game_flow/enum.h
@@ -69,7 +69,7 @@ typedef enum {
     GFS_SETUP_BACON_LARA,
 #if TR_VERSION == 2
     GFS_SET_START_ANIM,
-    GFS_ENABLE_SUNSET,
 #endif
+    GFS_ENABLE_SUNSET,
     GFS_NUMBER_OF,
 } GF_SEQUENCE_EVENT_TYPE;
